requirements.txt
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
mysqlclient==2.1.1
Pillow==9.2.0
requests==2.28.1
psycopg2==2.7.5
